HTTPS APIS -> BackendProject WebAPIS
This PR is to address the inablity for HTTPServer to run the other 3 APIs only seemingly available to HTTPS, even though we can edit minigames in development/testing to get these working. This PR addresses that so BOTH HTTPS **AND** HTTP Servers can run these 3 APIS properly.

Also adds more initial support for Homecafe minigames for PremiumAgency.
Adds InfoBoardSchedule.
