+++ BackendProject/WebAPIs/NDREAMS/Aurora/Blimp.cs
@@ -1,0 +1,7 @@
+﻿namespace BackendProject.WEBAPIs.NDREAMS.Aurora
+{
+    public class Blimp
+    {
+
+    }
+}
