+++ BackendProject/WebAPIs/NDREAMS/Aurora/OrbrunnerScores.cs
@@ -1,0 +1,7 @@
+﻿namespace BackendProject.WEBAPIs.NDREAMS.Aurora
+{
+    public class OrbrunnerScores
+    {
+
+    }
+}
