--- HTTPSecureServerLite/API/NDREAMS/Aurora/Blimp.cs
@@ -1,7 +1,0 @@
-﻿namespace HTTPSecureServerLite.API.NDREAMS.Aurora
-{
-    public class Blimp
-    {
-
-    }
-}
