--- HTTPSecureServerLite/API/NDREAMS/Aurora/OrbrunnerScores.cs
@@ -1,7 +1,0 @@
-﻿namespace HTTPSecureServerLite.API.NDREAMS.Aurora
-{
-    public class OrbrunnerScores
-    {
-
-    }
-}
