--- HTTPSecureServerLite/API/VEEMEE/Slot.cs
@@ -1,239 +1,0 @@
-﻿using BackendProject.MiscUtils;
-using CustomLogger;
-using HttpMultipartParser;
-
-namespace HTTPSecureServerLite.API.VEEMEE
-{
-    public class Slot
-    {
-        public static string? GetObjectSlot(byte[]? PostData, string? ContentType)
-        {
-            int max_slot = 0;
-            string slot_name = string.Empty;
-            string session_key = string.Empty;
-            string scene_id = string.Empty;
-            string region = string.Empty;
-            string object_id = string.Empty;
-            string psn_id = string.Empty;
-            string instance_id = string.Empty;
-            string hex = string.Empty;
-            string __salt = string.Empty;
-            string? boundary = HTTPUtils.ExtractBoundary(ContentType);
-
-            if (boundary != null && PostData != null)
-            {
-                using (MemoryStream ms = new(PostData))
-                {
-                    var data = MultipartFormDataParser.Parse(ms, boundary);
-
-                    slot_name = data.GetParameterValue("slot_name");
-
-                    session_key = data.GetParameterValue("session_key");
-
-                    scene_id = data.GetParameterValue("scene_id");
-
-                    region = data.GetParameterValue("region");
-
-                    try
-                    {
-                        max_slot = int.Parse(data.GetParameterValue("max_slot"));
-                    }
-                    catch (Exception)
-                    {
-                        // Not Important
-                    }
-
-                    object_id = data.GetParameterValue("object_id");
-
-                    psn_id = data.GetParameterValue("psn_id");
-
-                    instance_id = data.GetParameterValue("instance_id");
-
-                    hex = data.GetParameterValue("hex");
-
-                    __salt = data.GetParameterValue("__salt");
-
-                    ms.Flush();
-                }
-
-                return Processor.sign($"{{\"slot\":{SlotManager.UpdateSlot($"{instance_id}_{slot_name}", 0, psn_id, false, max_slot)}}}");
-            }
-
-            return null;
-        }
-
-        public static string? RemoveSlot(byte[]? PostData, string? ContentType)
-        {
-            int slot_num = 0;
-            string slot_name = string.Empty;
-            string session_key = string.Empty;
-            string scene_id = string.Empty;
-            string region = string.Empty;
-            string object_id = string.Empty;
-            string psn_id = string.Empty;
-            string instance_id = string.Empty;
-            string hex = string.Empty;
-            string __salt = string.Empty;
-            string? boundary = HTTPUtils.ExtractBoundary(ContentType);
-
-            if (boundary != null && PostData != null)
-            {
-
-                using (MemoryStream ms = new(PostData))
-                {
-                    var data = MultipartFormDataParser.Parse(ms, boundary);
-
-                    slot_name = data.GetParameterValue("slot_name");
-
-                    session_key = data.GetParameterValue("session_key");
-
-                    scene_id = data.GetParameterValue("scene_id");
-
-                    region = data.GetParameterValue("region");
-
-                    object_id = data.GetParameterValue("object_id");
-
-                    psn_id = data.GetParameterValue("psn_id");
-
-                    instance_id = data.GetParameterValue("instance_id");
-
-                    hex = data.GetParameterValue("hex");
-
-                    __salt = data.GetParameterValue("__salt");
-
-                    ms.Flush();
-                }
-
-                return Processor.sign($"{{\"success\":{SlotManager.UpdateSlot($"{instance_id}_{slot_name}", slot_num, psn_id, true)}}}");
-            }
-
-            return null;
-        }
-
-        public static string? HeartBeat(byte[]? PostData, string? ContentType)
-        {
-            string session_key = string.Empty;
-            string scene_id = string.Empty;
-            string region = string.Empty;
-            string slot_name = string.Empty;
-            string object_id = string.Empty;
-            string psn_id = string.Empty;
-            string instance_id = string.Empty;
-            string hex = string.Empty;
-            string __salt = string.Empty;
-            string? boundary = HTTPUtils.ExtractBoundary(ContentType);
-
-            if (boundary != null && PostData != null)
-            {
-                using (MemoryStream ms = new(PostData))
-                {
-                    var data = MultipartFormDataParser.Parse(ms, boundary);
-
-                    slot_name = data.GetParameterValue("slot_name");
-
-                    session_key = data.GetParameterValue("session_key");
-
-                    scene_id = data.GetParameterValue("scene_id");
-
-                    region = data.GetParameterValue("region");
-
-                    object_id = data.GetParameterValue("object_id");
-
-                    psn_id = data.GetParameterValue("psn_id");
-
-                    instance_id = data.GetParameterValue("instance_id");
-
-                    hex = data.GetParameterValue("hex");
-
-                    __salt = data.GetParameterValue("__salt");
-
-                    ms.Flush();
-                }
-
-                return Processor.sign("{ \"heartbeat\": true }");
-            }
-
-            return null;
-        }
-    }
-
-    public static class SlotManager
-    {
-        private static Dictionary<string, Dictionary<int, string>> instanceData = new Dictionary<string, Dictionary<int, string>>();
-
-        public static string UpdateSlot(string instance_id, int slot_num, string psn_id, bool removemode, int max_slot = 0)
-        {
-            bool found = false;
-
-            try
-            {
-                if (!instanceData.ContainsKey(instance_id))
-                {
-                    instanceData[instance_id] = new Dictionary<int, string>();
-
-                    // Initialize the dictionary with max_slot number of slots.
-                    for (int i = 1; i <= max_slot; i++)
-                    {
-                        instanceData[instance_id][i] = "<unnocupied/>";
-                    }
-                }
-
-                var data = instanceData[instance_id];
-
-                if (slot_num != 0 && removemode)
-                {
-                    if (data.ContainsKey(slot_num))
-                    {
-                        if (data[slot_num] == psn_id)
-                        {
-                            data[slot_num] = "<unnocupied/>";
-                            found = true;
-                        }
-                    }
-                }
-
-                if (slot_num == 0)
-                {
-                    foreach (var kvp in data)
-                    {
-                        if (!removemode)
-                        {
-                            if (kvp.Value == psn_id)
-                                return kvp.Key.ToString();
-
-                            if (kvp.Value == "<unnocupied/>")
-                            {
-                                data[kvp.Key] = psn_id;
-                                return kvp.Key.ToString();
-                            }
-                        }
-                        else
-                        {
-                            if (kvp.Value == psn_id)
-                            {
-                                data[kvp.Key] = "<unnocupied/>";
-                                found = true;
-                            }
-                        }
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                LoggerAccessor.LogWarn($"[VEEMEESlotManager] - Failed to update or remove slot - {ex}");
-
-                found = false;
-            }
-
-            if (!found)
-            {
-                if (!removemode)
-                    return "0";
-                else
-                    return "false";
-            }
-            else
-                return "true";
-        }
-    }
-}
